+++ app.py
@@ -17,6 +17,8 @@ ROUTE_MAP = [(re.compile(r'/[^/]*\.git.*'), git_http),
              (re.compile(r'/logout'), django_app),
              (re.compile(r'/mirrors'), django_app),
              (re.compile(r'/badge'), django_app),
+             (re.compile(r'/watch'), django_app),
+             (re.compile(r'/watching'), django_app),
              (re.compile(r'/vilya'), django_app),
              (re.compile(r'/.*'), web)]
